Micro refactor
