+++ src/app/recursos/__init__.py
@@ -1,0 +1,5 @@
+from .ping import Ping
+from .video import *
+from .usuario import *
+from .sesion import Sesion
+from .base_de_datos import BaseDeDatosResource
