+++ src/app/recursos/ping.py
