+++ src/app/recursos/usuario/__init__.py
@@ -1,0 +1,2 @@
+from .usuario import UsuarioResource
+from .perfil import PerfilUsuarioResource
