+++ src/app/recursos/usuario/perfil.py
@@ -3,31 +3,13 @@ from flask import request, g
 import auth_server_api
 from app.login_requerido_decorator import login_requerido
 
-class UsuarioResource(Resource):
-    def post(self):
-        post_data = request.get_json()
-        email = post_data['email']
-        password = post_data['password']
-
-        response = auth_server_api.registro_nuevo_usuario(email, password)
-
-        return response.json(), response.status_code
-
+class PerfilUsuarioResource(Resource):
     @login_requerido
     def get(self, usuario_id=None):
-        if not usuario_id:
+        if usuario_id is None:
             usuario_id = g.usuario_actual
-
-        response = auth_server_api.get_usuario(usuario_id)
-
-        return response.json(), response.status_code
-
-class PerfilUsuarioResource(Resource):
-
-    @login_requerido
-    def get(self):
-        response = auth_server_api.get_perfil(g.usuario_actual)
-
+        
+        response = auth_server_api.get_perfil(usuario_id)
         return response.json(), response.status_code
 
     @login_requerido
@@ -47,13 +29,5 @@ class PerfilUsuarioResource(Resource):
 
         response = auth_server_api.actualizar_perfil_usuario(nombre, apellido, telefono, direccion,
                                                              foto)
-
-        return response.json(), response.status_code
-
-class PerfilOtroUsuarioResource(Resource):
-
-    @login_requerido
-    def get(self, usuario_id):
-        response = auth_server_api.get_perfil(usuario_id)
 
         return response.json(), response.status_code
