--- src/app/resources/__init__.py
@@ -1,7 +1,0 @@
-from .ping import Ping
-from .video import VideoResource, VideoIdResource, VideoReaccion, VideoComentario
-from .usuario import UsuarioResource
-from .usuario import PerfilUsuarioResource
-from .usuario import PerfilOtroUsuarioResource
-from .sesion import Sesion
-from .base_de_datos import BaseDeDatosResource
